Implement console script for setup
