+++ setup.py
@@ -45,7 +45,9 @@ setuptools.setup(
     long_description_content_type = 'text/markdown',
     zip_safe = False,
     entry_points = {
-        'console_scripts': cfg.get('console_scripts','').split(),
+        'console_scripts': cfg.get('console_scripts','').split() + [
+            'pisces_setup = pisces.pisces_setup:pisces_setup',
+        ],
         'nbdev': [f'{cfg.get("lib_path")}={cfg.get("lib_path")}._modidx:d']
     },
     **setup_cfg)
